remove unofficial testnet dai
Official rinkeby support for MCD has been deprecated a long time ago. Furthermore the rinkeby address in this library is not even a dai.sol implementation - not sure who deployed or maintains this instance.
Instead I would encourage users to utilize the official goerli MCD instance, and the related Dai token for testing, which I have added in this commit.